Removed the *device* module, and updated the xspress module to not need ``make_device()``.
